refactor: extract connection and migration mixins from postgres backend
